pyproject.toml
[tool.poetry]
name = "danbibackend"
version = "0.1.0"
description = ""
authors = ["KimDuHong <art970@naver.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
Django = "^4.2"
drf-yasg = "^1.21.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
